Extract deploy stage cache support
